Fixes #11236: Windows port detection
This reverts commit 81553263ab812a7fd0a2ab0f627bee139ad6397c.

This fixes a regression with Windows port detection which led to port
collisions not being fixed on `vagrant up`.

PR #8517 changed `IsPortOpen#is_port_open?` to rescue
Errno::EADDRNOTAVAIL, but when we merged it into master, there was code
in `HandleForwardedPortCollisions#port_check` that depended on that
error bubbling up.

This also changes the port number in the HandleForwardedPortCollisions
tests to be a less common port number -- some of these tests will fail
if there is a local process running on port 8080.