Remove Windows GEMRC fix/workaround
The RubyGems pull-request (https://github.com/rubygems/rubygems/pull/2735) this temporary fix/workaround was waiting for to land, got merged in April of 2019.

The workaround is causing Vagrant to crash for many of our Windows users when they try to run `vagrant up` through and NPM-script. Removing the workaround code appears to solve their issue.