Detail other valid values for `config.ssh.verify_host_key`.
This commit adds more detail to the documentation for `config.ssh.verify_host_key`. Currently, the documentation only mentions the default value but does not provide any hints as to what valid values exist, or where this value is being passed to. As I'm sure you know, this is ultimately passed to the `net-ssh` library. It would have saved me some time doing searches to have read this fact in the Vagrant docs directly.

Thanks for considering this expansion.