save the mountpoint within virtualbox folder share
As of version 6.0 VBoxManage accepts the commandline argument "--auto-mount-point" (even though it is still missing in their documentation) which can be used to define the guestpath for a folder share. This has the advantage that mountpoints stay consistent,  say throughout a reboot procedure initiated from within the virtual machine or when starting the VM not via "vagrant up" but from within the VirtualBox Manager.
fixes hashicorp/vagrant#11415