Fix #8704: Double-check guest IP if it ends in .1
If the VirtualBox guest property `/VirtualBox/GuestInfo/Net/1/V4/IP`
returns an IP address ending in .1, assume it's a fluke and check again.

This fixes an issue that was revealed as an NFS error, where Vagrant was
creating an exports file with the wrong IP address. This was thought to
be caused by the presence of a docker0 interface, but it manifested
itself even without Docker installed.