Docker provider: bridge_inhibitipv4 option to prevent IP assignment
Make the scoped option docker_network__bridge_inhibitipv4 pass an argument to
`docker network create` that inhibits IP address assignemnt to the bridge:

    -o com.docker.network.bridge.inhibit_ipv4=true

This patch allows the upcoming Docker functionality to be access using a
network option. In future there might be a case for switching this to true by
default for private: networks for parity with other providers.

Rationale and context is provided in issue #11577.