Fixed error in bootstrap.sh
The previous version of the provisioning instructions linked to the `/vagrant` directory, but
the current version of Apache uses a default `DocumentRoot` setting of `/var/www`, not
`/var/www/html`, as assumed by the previous version of the documentation. The current version
of this tutorial results in the `/vagrant` directory being completely wiped out and replaced
by a new -and empty- `/vagrant` directory, by virtue of the `--force` flag specified by the
documentation.