Don't try to "fix" a /etc/fstab that does not exist
remove_vagrant_managed_fstab() gets called in self.persist_mount_shared_folder
when folders is nil. This is an issue for boxes that don't have a /etc/fstab, as
they can boot but will break once the sed call is executed.
Thus we guard it with a check whether /etc/fstab exists and only try to remove
the modifications then.