Pre-Commit Hook for Vagrant Validate
Git pre-commit hooks are a great integration point for linters, because
they (1) guarantee that all committed code is consistently formatted and
syntactically correct and (2) do not require you to run linters on every
build or test. Since Vagrant ships with a `validate` command, it is, in
a way, an elaborate linter and so it makes sense to integrate this
linting functionality into the Git lifecycle. Support for Pre-Commit, a
"framework for managing and maintaining multi-language pre-commit
hooks," is added through the addition of the .pre-commit-hooks.yaml
file.