Do not rely on hardcoded PKG_PATH URI but rather on the default behavior of pkg_add(1) which selects the mirror based on the content of '/etc/installurl' to install rsync on OpenBSD targets.
This behaviour is present since OpenBSD 6.1 (released April 11, 2017).

The hardcoded URI prevents the use of custom OpenBSD mirrors (written down to '/etc/installurl' by the installer) while using the rsync directive in Vagrantfiles.

This prevents 'Installing rsync to the VM...' failure when provisioning a VM configured with a custom mirror.

https://man.openbsd.org/pkg_add.1
https://man.openbsd.org/installurl.5