Tests for the basis service
Filling these out as a part of a sweep to ensure we return decent error
messages from our GRPC services.

For this service I didn't come up with any error condition that
warranted changing the code yet. The get -> notfound case seems to be
the main one and the pass through error object seems AOK here.