Homepage: mention that Vagrant is free software
As far as I can tell Vagrant is free software. This is a big deal for some people (like me). This is a big enough feature to mention it on your homepage.

Note: I could not find a detailed list of components/plugins and their licenses in your documentation. Is all of Vagrant MIT-licensed? Are some parts released under a different license?